refactor(frontend): extract quote charge card
